feat: implement submit party functionality with validation and storage
- Add submitParty.ts for handling party-related logic including default weapon selection and compatibility checks.
- Introduce submitSteps.tsx as a placeholder for future step components.
- Create submitStorage.ts for managing draft state persistence and merging.
- Implement submitTime.ts for time input parsing and formatting.
- Develop submitUi.tsx for UI components related to the submission process.
- Add submitUid.ts for handling character selection and UID management.
- Create submitValidation.ts for validating submission steps and ensuring data integrity.
- Define types in types.ts for submission draft and party slots.
